refactor(client): unify dashboard + analysis-builder filter logic into one module (#913)
Consolidate the duplicated filter operator/manipulation logic into a single
deep filter module at src/client/shared/filters, consumed by both the
dashboard filter UI and the analysis-builder filter UI.

- New shared/filters module: operators.ts (FILTER_OPERATORS + getAvailableOperators),
  filterOperations.ts (type guards, creators, flatten/count/extractFilterMembers,
  addFilterAtPath/removeFilterAtIndex/toggleGroupType, findDateFilterForField/
  removeFilterForMember, server transforms, validateFilterForCube), index.ts barrel.
- AnalysisBuilder: delete inline guards/countFilters/getSelectedFields/addFilterAtPath
  and the AND/OR toggle+unwrap logic; components consume the module. Comparison
  helpers delegate to the module's findDateFilterForField/removeFilterForMember.
- Dashboard utils/filterUtils: delete convertToServerFormat (-> transformFiltersForServer),
  extractMemberNamesFromFilter + extractFieldsFromFilters (-> extractFilterMembers),
  and the duplicate validateFilterForCube (now imported + re-exported).
- Move FILTER_OPERATORS/FilterOperatorMeta out of shared/types.ts and the filter
  helpers out of shared/utils.ts; all barrels (shared/index, shared/types,
  components/shared/utils) keep their existing export names, sourced from ./filters.
- Unify isSimpleFilter to require member+operator (not values) so valueless
  inDateRange filters are handled consistently across both surfaces.
- Add tests/client/filterOperations.test.ts covering the module directly;
  migrate the validateFilterForCube tests out of filterUtils.test.ts.

typecheck, test:client (5893), lint, build, check:exports all green.
