+++ src/client/components/DashboardFilters/useDashboardFilterConfigModal.ts
@@ -20,8 +20,8 @@ import { useTranslation } from '../../hooks/useTranslation.js'
 import type { DashboardFilter, SimpleFilter, FilterOperator } from '../../types.js'
 import type { MetaResponse, MetaField } from '../../shared/types.js'
 import type { FieldType } from '../AnalysisBuilder/types.js'
-import { FILTER_OPERATORS, DATE_RANGE_OPTIONS } from '../../shared/types.js'
-import { getAvailableOperators } from '../../shared/utils.js'
+import { DATE_RANGE_OPTIONS } from '../../shared/types.js'
+import { FILTER_OPERATORS, getAvailableOperators } from '../../shared/filters/index.js'
 import { findFieldInSchema, getFieldTitle } from '../AnalysisBuilder/utils/index.js'
 import { computeDirectInputValues } from './dashboardFilterConfigModalUtils.js'
 import { useFilterDropdowns } from './useFilterDropdowns.js'
